requirements.txt
beautifulsoup4==4.9.3
et-xmlfile==1.0.1
openpyxl==3.0.7
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
selenium==3.141.0
urllib3==1.26.4
Pillow==8.2.0
pytesseract==0.3.7